fix: build 완료

+++ src/main/java/com/codeit/deokhugam/domain/entity/Book.java
@@ -39,7 +39,7 @@ public class Book extends BaseUpdatableEntity {
 
   @Builder.Default
   @Column(name = "review_count", nullable = false)
-  private int reviewCount = 0;
+  private Long reviewCount = 0L;
 
   @Builder.Default
   @Column(name = "rating", nullable = false, precision = 3, scale = 2)
